Rename `netusers` to `schema` to be more generic
